Refactor fuzz_glsl/spirv_test files in gfauto
Refactors the files that take care of GLSL and SPIR-V fuzzing in
gfauto to (a) make them Amber-specific (paving the way for new
ShaderTrap-related files in the future), and (b) prevent them from
depending on one another by extracting common functionality to a
utilty file.

Also fixes a minor style issue where a variable was named "match",
which was leading to a warning in PyCharm.

Related issue: #1160.
